requirements.txt
python-dotenv==1.0.1
requests==2.32.3
openai>=1.2.3
slack_bolt==1.23.0
requests-toolbelt==1.0.0